requirements.txt
# Core Data Science
numpy==1.24.3
pandas==2.0.3
pyarrow==12.0.1
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0

# Explainability
shap==0.42.1
dice-ml==0.10

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Configuration & Validation
pyyaml==6.0.1
pydantic==2.1.1
python-dotenv==1.0.0

# Calendar & Time
holidays==0.29
pytz==2023.3

# HTTP & API
requests==2.31.0
urllib3==2.0.4

# Jupyter
jupyter==1.0.0
ipykernel==6.25.0
ipywidgets==8.1.0

# Utilities
tqdm==4.65.0
joblib==1.3.1
click==8.1.7

# API
fastapi==0.109.0
uvicorn==0.27.0
pydantic-settings==2.1.0

# Optimization
optuna==3.5.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Documentation
Sphinx==7.1.2
